requirements.txt
opencensus-ext-azure
python-dotenv
faker